# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ledgerline API
  version: 1.0.0
  description: Multi-tenant double-entry accounting API.
servers:
  - url: http://localhost:3000/api/v1
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            requestId:
              type: string
          required: [code, message]
      required: [error]
...
    InvoiceLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lineNo:
          type: integer
        description:
          type: string
        accountId:
          type: string
          format: uuid
        quantity:
          type: string
          example: "2.000000"
        unitPrice:
          type: string
          example: "500.0000"
        discountPct:
          type: string
          example: "0.00"
        taxCodeId:
          type: string
          format: uuid
        taxAmount:
          type: string
          example: "130.0000"
        lineTotal:
          type: string
          example: "1130.0000"
      required: [lineNo, description, accountId, quantity, unitPrice, lineTotal]
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        docNo:
          type: string
        docDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        partyId:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, posted, paid, partially_paid, reversed]
        subtotal:
          type: string
        discountAmount:
          type: string
        taxableAmount:
          type: string
        taxAmount:
          type: string
        grandTotal:
          type: string
        outstandingAmount:
          type: string
        version:
          type: integer
        journalEntryId:
          type: string
          format: uuid
          nullable: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLine'
      required: [id, docNo, docDate, partyId, status, grandTotal, outstandingAmount, version, lines]
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    OrganizationHeader:
      name: X-Organization-Id
      in: header
      required: true
      description: Which organization this request operates on. Verified server-side against an active membership.
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /invoices:
    get:
      summary: List invoices
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - name: partyId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, posted, paid, partially_paid, reversed]
        - name: from
          in: query
          required: false
          description: Inclusive document-date lower bound.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          description: Inclusive document-date upper bound.
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
    post:
      summary: Create a draft invoice
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                partyId:
                  type: string
                  format: uuid
                docDate:
                  type: string
                  format: date
                dueDate:
                  type: string
                  format: date
                lines:
                  type: array
                  items:
                    $ref: '#/components/schemas/InvoiceLine'
              required: [partyId, docDate, lines]
      responses:
        '201':
          description: Draft invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
  /invoices/preview:
    post:
      summary: Server-computed totals for an in-progress invoice, before saving
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lines:
                  type: array
                  items:
                    $ref: '#/components/schemas/InvoiceLine'
              required: [lines]
      responses:
        '200':
          description: Computed subtotal, tax, and grand total — not persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
  /invoices/{id}:
    get:
      summary: Get one invoice, with its lines
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a draft invoice (posted invoices are immutable)
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: integer
                  description: Optimistic concurrency check — must match the current version
                partyId:
                  type: string
                  format: uuid
                docDate:
                  type: string
                  format: date
                lines:
                  type: array
                  items:
                    $ref: '#/components/schemas/InvoiceLine'
              required: [version]
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '409':
          description: Version mismatch — someone else edited this draft first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
  /invoices/{id}/post:
    post:
      summary: Post a draft invoice — creates its balanced journal entry
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invoice posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
...